Remove feature to set up constructor properties
This makes the generated code easier to read and the algorithm easier to read. Constructor properties aren't really used or required so there isn't much loss. Generally you will set up interfaces in tests, and then use a test adapter of some kind to instantiate the system under test, and this test adapter can pass the relevant thiings in to the constructor