End to End test during build by generating and runnint tests
The plan is to use the SampleTests as documentation and as executable End to End tests. We will regenerate
the tests from the Excel spreadsheet and then run them, which will ensure that the test generation works and
produces working code.

Later we plan to also regenerate the Excel spreadsheet when running the tests. If possible we will compare
the pre and post generated Excel sheet to make sure that they are the same. If this isn't possible we will artifact
them both so at least we can compare manually. We could also go round the process again to make sure that
the created excel sheet is viable