Add row between end of given and start of when
When generating the c# code from the excel tests, "When" can
start immediately after the "Given" section

However when generating the excel from C#, a blank row is
inserted.

So in order to make the test round trippable I have added a
blank row to the spreadsheet.

It would be better if there was an error / warning about this,
probably when creating the C# tests, but possibly in both
directions.