Improve end of table detection when converting excel to code
Before the entire row had to be blank, or a preceding column
had to have a value. This wasn't working for the Anova example
as there were some totals in the row. These were part of the
excel calculations, rather than part of the test.

The logic now only states that all columns in the row up to the
end of the table must be blank in order to detect the end of a
table