Avoid unecessary rewriting of cells when converting code to excel
Overwriting cells should be avoided as much as possible, in case
they have formulas (which aren't overwritten, and hence cause
trouble).

The code was overwriting cells unecessarily, due to rounding errors,
and due to the excel to c# naming conversions. The comnparision
is now done on the raw Object value and on the value formatted as
a string