Fix setup of complex property within a table when at end
The detection of the end of the sub property wasn't working
properly, when there was no following column.

Also removed some code to detect the start column. This used to
be required when there were creational properties and normal
properties (which was necessary when the framework was setting up
real classes, but is not required when using mocks / interfaces)