Restore full build process
The build was failing because one of the sample tests was not round
trippable, but was attempting to round trip. I thought that the
code to stop it doing this was already written, but it was not. It
has been added now, so the full build process should be working
again.