Improve equality assertion
It used to return false for an int 3 being compared to a double 3,
whereas now this returns true.

There are still potential problems with round errors and
unrepresentable numbers and suchlike, but it is a lot better, and
you can always use "Percentage Precision" to allow a range of values
if needs be.