Detect when "With Properties" cell is missing for a table
A warning will be added to the console output when generating the
tests, and an error code returned.

Instead of producing un-compilable code, the problem worksheet will
instead create a file with a single comment explaining the problem.

This commit also outputs cell references in warning / error messages
in A1 format instead or R1C1 format.