Introduce a 'List of' syntax for setting up lists
This commit allows the syntax to be converted from excel to c#, but
the reverse does not yet happen, nor is a warning output to say that
this is the case.

The old / existing style of setting up lists is still present, but
will be removed later, which is a breaking change and will require
an change to most of the end to end tests.

This new syntax will make it more obvious when a list property, as
opposed to a single property is being set up, and will make the
conversion from code back to excel easier (properties such as
isChild and indexInParent will no longer be required).

This commit continues the current (bad) convention of having too
much code inside ExcelToCode. I plan to split out the 'list of'
code later, and then maybe split out other bits of functionality,
using this as the model.