Start to test output of properties using "list of" syntax
This compiles, and the tests passes, but it doesn't assert anything
yet

Need to fill in WriteClass of RunSpecification to handle the new
property on IReportsSpecificationSetup.

There might also be a failing test, locally nunit is trying to run
the test on Specification Base, which is abstract and so doesn't
work. This didn't used to happen so hopefully its just a problem
with my local setup.