Record complex properties
The plan is to record all given classes, and then use this to
generate the SpecificationSpecific classes where possible, by
searching for matching interfaces in an assembly. Simple and complex
properties are now done, table and list properties still need doing.

This commit changes the visitors (and some other bits of code) to
store / notify the Excel property names rather than the c# ones.

This is basically required for automatically creating classes, and
I think is also a good idea generally. We can still convert in to
the c# representation whenever we want, and it hints at supporting
other languages, and at splitting up ExcelToCodeConverter in to
general parsing stuff and language specific stuff.