Record Table properties
The plan is to record all given classes, and then use this to
generate the SpecificationSpecific classes where possible, by
searching for matching interfaces in an assembly.

This commit completes the recording of the classes, now need to
use this to match against interfaces / classes of the system
under test and other assemblies.