Add list classes to SpecificationSpecific class generation
When setting up an interface via a mock, it doesn't make sense to
support functions (a function with no parameters and a return value
is like a property so could work, but anything with parameters
cannot)

So I probably need to update the matching code to not match on
these

It is possible to use single parameter functions if setting up an
object / class directly, instead of doing it through a mock /
interface, but I think this is a rare case, and I'm not planning to
support it in the class generation. Potentially I should remove
support for it full stop.

This commit also introduces a Liskov problem in IGivenClassProperty,
I need to think about what I should do about that.