Add error handling around assembly loading
It would probably be better to pass in the list of types to
TestProjectCreator, instead of having it create them itself from the
list of filenames.

Ideally I wouldn't put this code in program.cs though, so I would
probably want to introduce two new classes, one to load the types from
the assemblies, and one to handle the workflow of getting the types and
then calling TestProjectCreator.

I don't think this is warranted yet, but would be if more similar things
were useful.

Putting it in program.cs now and then refatoring later is also a
possibility