Support primtive lists
Lists of floats were already supported, but it turned out that they only
worked with integer values, as floating point literals are `double` in
C#. I just went with this and changed them to `double`. This probably
makes sense when thinking about supporting other languages, such as
python, as `float` is the more general name.