Recognise custom date formats, and drop support for TimeSpan
It is only possible to recognise a date value in excel from the number
format, which is utterly idiotic. Previously the code only recognised
built in date formats, but now it should support most custom formats
as well, although when round tripping for format will revert to a built
in one.

TimeSpan is very csharp specific, and thinking about extending this to
python. Its also easy to convert from a date time to a time span
yourself, so the framework isn't really adding much.