Move parsing of complex properties in to new class
Previously just the CanParse function was moved, but now virtually
eveything is in the new class.

This has created a lot of duplcation temporarily, most of which can be
removed once list parsing is also in its own class.

This is an early commit, so there is also a lot of general tidy up to do