Improve assembly loading
Assembly loading is very weird in csharp, but you basically need
to load dependent assemblies yourself as well, instead of expecting
the framework to do it.

Also it will now load as many types as it can, whereas before if any
type failed to load then the whole assembly failed to load.