Change order of properties that are added GiveClass
One of the sample tests implicitly relies on AddProperty being
called in the order defined in the spreadsheet (which is bad), and
has caused a test to start failing.

This is one of the problems I am trying to fix at the moment, but in
the meantime I can change to order of the properties that are added
which happens to coincide with the way that the sample test wants
it to happen.