Add System.IO.Packaging as a dependency and move to .net standard
System.IO.Packaging is a dependency of DocumentFormat.OpenXml, and
should be brought in by it, but it isn't for some reason. I think
because I am using a .nuspec file to pack, instead of the .csproj.

I can't use the .csproj though, as it doesn't support the <files>
element

So I'm just adding the dependency manually